requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.3.1
alembic==1.8.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
beartype==0.19.0
beautifulsoup4==4.11.1
bleach==5.0.1
blis==0.7.11
boltons==23.0.0
Bottleneck==1.3.5
brotlipy==0.7.0
catalogue==2.0.10
certifi==2024.2.2
certipy==0.1.3
cffi==1.15.1
charset-normalizer==2.1.1
click==7.1.2
cloudpathlib==0.16.0
colorama==0.4.6
commonmark==0.9.1
conda==23.3.1
conda-package-handling==2.0.2
conda_package_streaming==0.7.0
confection==0.1.4
conllu==4.5.3
contourpy==1.0.5
cryptography==38.0.4
cupy==10.1.0
cycler==0.11.0
cymem==2.0.8
Cython==0.29.30
datasets==2.20.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
docopt==0.6.2
en-core-web-sm==3.7.1
en-ner-bc5cdr-md==0.5.4
entrypoints==0.4
et-xmlfile==1.1.0
evaluate==0.4.0
exceptiongroup==1.1.3
executing==0.10.0
fastjsonschema==2.16.1
fastrlock==0.8
feret==1.3.1
filelock==3.9.0
flit_core==3.6.0
fonttools==4.34.4
frozenlist==1.3.3
fsspec==2024.5.0
gdcm==1.1
greenlet==1.1.2
huggingface-hub==0.24.5
humanize==4.4.0
hydra-core==1.3.2
idna==3.6
imageio==2.19.3
imbalanced-learn==0.10.1
imblearn==0.0
ipykernel==6.15.1
ipython==8.17.2
ipython-genutils==0.2.0
ipywidgets==8.0.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.7.2
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-telemetry==0.1.0
jupyterhub==2.3.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
kiwisolver==1.4.3
langcodes==3.4.0
language_data==1.2.0
leveldb==0.201
lightning-utilities==0.12.0
lxml==4.9.1
Mako==1.2.1
marisa-trie==1.1.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
monai==1.4.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.10
mysql-connector-python==8.0.30
narwhals==1.29.1
nbclient==0.6.6
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.4
nibabel==5.0.0
nltk==3.8.1
nmslib==2.1.1
notebook==6.4.12
nptyping==2.3.1
numexpr==2.8.4
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
omegaconf==2.3.0
opencv-python==4.6.0.66
openpyxl==3.1.2
packaging==22.0
pamela==1.0.0
pandas==2.3.2
pandocfilters==1.5.0
parso==0.8.3
pathlib_abc==0.1.1
pathy==0.11.0
patsy==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pip==24.3.1
pluggy==1.0.0
preshed==3.0.9
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==17.0.0
pyarrow-hotfix==0.6
pybind11==2.6.1
pycosat==0.6.4
pycparser==2.21
pydantic==1.7.4
pydantic_core==2.18.3
pydicom==2.3.0
Pygments==2.16.1
pykwalify==1.8.0
pylibjpeg==1.4.0
pylibjpeg-libjpeg==1.3.2
pynrrd==1.0.0
pyOpenSSL==22.0.0
pyparsing==3.1.2
pyradiomics==3.0.1
pyrsistent==0.18.1
pysbd==0.3.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.4
pytorch-lightning==2.5.0
pytz==2022.7
PyWavelets==1.3.0
PyYAML==6.0
qtconsole==5.3.2
QtPy==2.2.0
quickumls==1.4.2
quickumls-simstring==1.1.5.post1
regex==2022.10.31
requests==2.32.3
responses==0.18.0
rich==12.6.0
rouge-score==0.1.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sacremoses==0.0.53
safetensors==0.4.0
scikit-image==0.19.3
scipy==1.8.1
scispacy==0.5.4
seaborn==0.12.1
Send2Trash==1.8.0
sentencepiece==0.1.97
setuptools==68.2.2
shellingham==1.5.0.post1
SimpleITK==2.2.0
six==1.16.0
smart-open==6.4.0
soupsieve==2.3.2.post1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.39
srsly==2.4.8
stack-data==0.5.0
statsmodels==0.14.2
sympy==1.13.1
terminado==0.15.0
thinc==8.2.3
threadpoolctl==3.1.0
tifffile==2022.5.4
tinycss2==1.1.1
tokenizers==0.14.1
toolz==0.12.0
torch==2.6.0
torchaudio==0.13.0
torchio==0.20.1
torchmetrics==1.6.1
torchvision==0.14.0
tornado==6.2
tqdm==4.66.5
traitlets==5.3.0
transformers==4.35.0
triton==3.2.0
typer==0.3.2
typing_extensions==4.12.0
tzdata==2025.2
Unidecode==1.3.8
unqlite==0.9.6
urllib3==1.26.18
wasabi==0.10.1
wcwidth==0.2.5
weasel==0.3.4
webencodings==0.5.1
wheel==0.41.2
widgetsnbextension==4.0.2
wrapt==1.14.1
xxhash==3.2.0
yarl==1.8.2
zstandard==0.19.0
